Refactored to idStrategy namespace
